fix imports problems in python3
